Initial checkin of Cell Broadcast display application.
Simple Cell Broadcast application. Supports ETWS and CMAS emergency
messages, and can be extended to support other types of PWS
(Public Warning System) emergency warnings. User preferences for
configuration of each supported national system (ETWS, CMAS, and
channel 50 broadcasts for Brazil) are displayed/hidden based on the
values in res/values/config.xml.

For emergency alerts, a flashing warning icon and special alert sound
is played. If there is a text-to-speech engine installed for the
language of the broadcast, then the contents of the broadcast message
will be spoken after the alert sound is played. The user can disable
the text-to-speech feature in settings, as well as adjusting the
length of the alert sound and enabling/disabling delivery of several
emergency broadcast channels.

The CellBroadcastReceiverTests apk includes a launcher activity with
buttons for sending each type of test message. This APK must be signed
with the system certificate in order to acquire the BROADCAST_SMS
permission required to send test broadcasts to the app.

Change-Id: If9b61e02c246de5783b3e39cd100ea707ea80084