Add ContentProvider for received cell broadcasts.
Enable apps with android.permission.READ_CELL_BROADCASTS to access the
database of received cell broadcasts via a ContentProvider at URI
"content://cellbroadcasts/". The ContentProvider is read-only, but
provides internal methods for the app to insert and delete broadcasts,
and to mark broadcasts as read. These methods are not exposed to other
applications.

Other changes:

 * Added static log() methods for simpler debug logging.

 * Moved CellBroadcastDatabase.DatabaseHelper to new top-level
   CellBroadcastDatabaseHelper class after refactoring the original
   CellBroadcastDatabase class out of existence.

 * Removed the "Delete" dialog title per Android UI guidelines,
   and changed from a custom view to a built-in dialog type.

Bug: 5752924
Change-Id: I1ad6a4d5248f851c53a638c2cb7e85a57ba6b811