Add "View details" context menu to CMAS alert list view.
For CMAS cell broadcasts, show the CMAS details (alert category,
response type, severity, urgency, certainty) in a context menu
accessed by long-pressing the alert and selecting "View details".
The CMAS alert details are no longer shown in the alert dialog
when a broadcast is received or short-pressed in the list view.

Also changed the CMAS detail headings and values to match the
strings in TIA-1149-0-1 (CMAS over CDMA systems) and removed the
trailing spaces from the heading string resources because they
are stripped out by aapt. Instead, we append a space between
the headings and the values.

For i18n purposes, we may want to consider switching to using a
format string for the headings so that the values can be placed
before, after, or in the middle of the heading.

Bug: 7041437
Change-Id: I18c40b14d8dcb811611c1da81e1cd2f6f51a5aa4