Enable support for area info cell broadcasts in Brazil.
Brazilian operators require support for displaying the most recent
area info cell broadcast (message identifier == 50) received.
Handle this with a broadcast that the Settings app can receive
and display in the status screen. Settings can also request
CellBroadcastReceiver to send the most recent area info received.

Also, when the SIM country code is Brazil, show the checkbox to
enable/disable channel 50 in cell broadcast settings, and enable
channel 50 in the radio at startup unless the user disables it.

Also update the test application to allow message identifier to be
set for GSM and UMTS cell broadcasts (excluding ETWS) for testing
the handling of different message identifiers, and change the
message identifier of the test cell broadcast PDUs from 50 to 1.
To override the message identifier, set the category field in the
test app to a non-zero value.

Bug: 7445881
Change-Id: Ibca8d33c32c0391d0d4eee444c86dd93878d7d1e