CBS : Update the "Delete All" display.
In the CellBroadcastReceiver application, when the user
presses for the first time the menu touch, with one or
more SMS_CB received already, the "Delete All" is not
displayed.

This patch updates the menu display of the
CellBroadcastReceiver application. With this patch,
the "delete all" appears correctly in the settings
menu, if one or several SMS are present in the application.

Change-Id: I7bc6397ce4f024378cd964a44c6ec2a8f3bdbbf4
Author: Emmanuel Delaude <emmanuelx.delaude@intel.com>
Signed-off-by: Emmanuel Delaude <emmanuelx.delaude@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 65376