Add CMAS reminder alert feature.
Carrier requirement: add "alert reminder" feature for CMAS alerts.
If the alert is not dismissed, a reminder sound is played based on
user preference. The user can select one of the following behaviors:
 - "Once" = one alert reminder after 2 minutes
 - "Every 2 minutes"
 - "Every 15 minutes"
 - "Off" = no alert reminders

The notification sound is the user's preferred notification sound, or
the user's preferred alarm sound as a fallback.

Bug: 7045215
Change-Id: Id7858a11dd9b2227aea8ea4a826ed0e78c1ffcb6