Fix overlapping text in CMAS list items.
If the font is set to a large size, CMAS alerts with long titles
such as "Child abduction (Amber alert)" will draw on top of the
date. Fix the layout to ensure the title is left-justified
and ellipsized to fit on one line if necessary. Tested with all
four standard font sizes in both portrait and landscape mode.

Bug: 10858267
Change-Id: I871a924d516ae29da180280e7579ff89cffdd22a