Fixed that CMAS duplicate detection is not working
Now we'll have different duplicate detection algorithm between
ETWS and CMAS message.

If this is an ETWS message, then we want to include the body message
to be a factor for duplicate detection. We found that some Japanese
carriers send ETWS messages with the same serial number,
therefore the subsequent messages were all ignored.

In the other hand, US carriers have the CMAS requirement that only
serial number, location, and category should be used for duplicate
detection.

bug: 22787502
Change-Id: I92ef2d47a76622ee1be0d23ae26bf20a1dcaf203