Fixed that emergency alert reminder does not trigger every 2 minutes.
The alert was not triggered because the device was in deep sleep. The
OS shifts the alarms in order to minimize wakeups and battery use. Since
this is emergency scenario, we use AlarmManager.setExact to trigger
the alarm in every 2 or 15 minutes precisely.

bug: 24391040
Change-Id: I7a3812b48543a6e4dfc0efe4791d21da9feab771
(cherry picked from commit 2c71594ba2d74e37a2e80948c3bb36150ff9ad6d)