Fixed that sometimes ETWS messages are not spoken.
Now we will use the language specified in ETWS/CMAS messages. If
the language is not specified in ETWS, or the device does not support
that language, we use Japanese for ETWS. If the language is not
specified in CMAS, or the device does not support that language, we
use device's default language for CMAS.

Also we found that sometimes ETWS primary and secondary messages come
back to back. Also there should not be any text in the primary message
so there is no point to play TTS for that. So we disable the TTS
for ETWS primary message and only let the secondary one played out.
Also change the TTS parameter from QUEUE_FLUSH (only flush the queue
from the same caller) to QUEUE_DESTROY (flush all queues from all
caller) since this is for emergency.

bug: 24341917
Change-Id: I431d2d8b1e99130ff5e1d1931992b63cf6925778