Support CMAS regulatory requirements
Fixes an issue where the 10.5s CMAS alert audio can be truncated
because the code relied on a timer and assumed playback would
start with zero lag. The alert audio file is changed to include
the full 10.5s alert sequence and is played in full without a
timer if the requested duration is 10.5s.

Change-Id: I0fff72579cad74bdb0ef5ab2fe6be8daf5c3b837
BZ: 25728174