Removed emergency tone duration configuration from the UI
Since CMAS uses fixed 10.5 seconds alert tone, and most of
the carriers using ETWS also have fixed length tones, the
tone duration configuration UI might confuse users. Since
this is just for testing purpose, which is only shown to
the user when developer option is turned on, we decided to
remove it.

bug: 28563429
Change-Id: I227fedb73816f7c8026238eee384b52a1d1e684f