Configure the channel for customized UI
Add a new customized setting UI to support channel configure.
User could disable any CB channel by adding it to the UI list
and keep it unchecked. If the customized channel is checked,
the channel will be enabled, and if any channel is not listed,
it would be configured by default.

Change-Id: Iab2333a797774ffaad2421f5346db4d6d460104f
CRs-Fixed: 1037633