CellBroadcastReceiver: Show timestamp / appropriate title in the LatAm region
In this region, broadcasts on the 4370 are emergency alerts. 4370 is
also the channel used for presidential alerts on the US, so piggy back
on that code and change the string for LatAm.

Additionally, the timestamp is mandatory in some government
requirements, this patch adds that as well.

Change-Id: Ifc5ef92bc97551d69102a915df447813cfa8395a
Ticket: FEIJAO-1615
(cherry picked from commit 60fd85ca6a06b9fdcc2ffcf120a692d90aa1d1ec)